Create Unit Details Screen (#53)
* Create Unit Details Screen

- Added a screen to the pause menu for displaying/managing units
- Currently the screen displays the units found by the data manager using a resource lookup
- Currently the actual unit details screen used to display the unit's information is quite limited

* Rearrange Modular Character transform hierarchy

Needed to be done eventually to support multiple different character
models, and the best time to do that is now before the prefab gets used
in multiple places (which the unit details screen will cause to happen)

* Add unit display model to unit details screen

Added the unit customizer prefab to the unit details screen so the unit
can be visually displayed. Had to manipulate the canvas a lot. Ended up
converting the canvas to a world space canvas and adding a second
overlay camera for the UI, but it works.

* Add unit inventory display to unit details screen

* Unit Inventory Management

In the Unit Details Screen, when a unit is selected, a menu opens so the
player can move items into and out of that unit's inventory